requirements.txt
apache-airflow==2.8.1
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
pytest==8.1.0